+++ release_notes.md
@@ -5,4 +5,4 @@
 
 - Add Retry Policy Attributes (#977, #971)
 - Bug fix - GetOutputBindings returns incorrect data when OutputBindingData is not set (#983)
-
+- Worker version and environment information returned in init call (WorkerMetadata)